refactor : assets 폴더 경로 변경
각 컴포넌트 폴더 생성 후 삽입, Button 컴포넌트 너비 반응 수정
