+++ src/components/Button/Button.tsx
@@ -3,7 +3,7 @@ import { ButtonProps } from 'types/Button.type';
 export const LargeButton = ({ text, bgColor, textColor = 'text-black' }: ButtonProps) => {
   return (
     <button
-      className={`flex justify-center items-center gap-2 px-min-24px py-4 h-[70px] w-[132px] border-txt_primary border-2 shadow-[0_4px_4px_rgba(0,0,0,1)] rounded-[16px] ${
+      className={`flex justify-center items-center gap-2 px-min-24px py-4 h-[70px] w-full max-w-full min-w-[142px] border-txt_primary border-2 shadow-[0_4px_4px_rgba(0,0,0,1)] rounded-[16px] ${
         bgColor ? bgColor : 'bg-main_primary'
       }`}
     >
@@ -17,7 +17,7 @@ export const LargeButton = ({ text, bgColor, textColor = 'text-black' }: ButtonP
 export const MediumButton = ({ text, bgColor, textColor = 'text-black' }: ButtonProps) => {
   return (
     <button
-      className={`flex justify-center items-center gap-2 px-min-20px py-4 h-[58px] w-[110px] border-txt_primary border-2 shadow-[0_4px_4px_rgba(0,0,0,1)] rounded-[16px] ${
+      className={`flex justify-center items-center gap-2 px-min-20px py-4 h-[58px] w-full max-w-full min-w-[110px] border-txt_primary border-2 shadow-[0_4px_4px_rgba(0,0,0,1)] rounded-[16px] ${
         bgColor ? bgColor : 'bg-main_primary'
       }`}
     >
@@ -31,7 +31,7 @@ export const MediumButton = ({ text, bgColor, textColor = 'text-black' }: Button
 export const SmallButton = ({ text, bgColor, textColor = 'text-black' }: ButtonProps) => {
   return (
     <button
-      className={`flex justify-center items-center gap-2 px-min-16px py-3 h-[46px] w-[88px] border-txt_primary border-2 shadow-[0_4px_4px_rgba(0,0,0,1)] rounded-[16px] ${
+      className={`flex justify-center items-center gap-2 px-min-16px py-3 h-[46px] w-full max-w-full min-w-[88px] border-txt_primary border-2 shadow-[0_4px_4px_rgba(0,0,0,1)] rounded-[16px] ${
         bgColor ? bgColor : 'bg-main_primary'
       }`}
     >
